Refactor UnassignedView and WorktreeChanges for rules component
This commit refactors the UnassignedView and WorktreeChanges components to support inline integration of the rules viewer. It introduces a foldable UI for the unassigned sidebar, updates props to support slot rendering, and cleans up references to previously handled fold button logic. These changes improve the extensibility and layout cohesion for the rules component.
